+++ core/src/banking_stage/transaction_scheduler/prio_graph_scheduler.rs
@@ -1,6 +1,7 @@
 use {
     super::{
         in_flight_tracker::InFlightTracker,
+        scheduler::Scheduler,
         scheduler_error::SchedulerError,
         thread_aware_account_locks::{ThreadAwareAccountLocks, ThreadId, ThreadSet},
         transaction_state::SanitizedTransactionTTL,
@@ -13,7 +14,8 @@ use {
             ConsumeWork, FinishedConsumeWork, MaxAge, TransactionBatchId, TransactionId,
         },
         transaction_scheduler::{
-            transaction_priority_id::TransactionPriorityId, transaction_state::TransactionState,
+            scheduler::SchedulingSummary, transaction_priority_id::TransactionPriorityId,
+            transaction_state::TransactionState,
         },
     },
     crossbeam_channel::{Receiver, Sender, TryRecvError},
@@ -63,7 +65,9 @@ impl PrioGraphScheduler {
             prio_graph: PrioGraph::new(passthrough_priority),
         }
     }
+}
 
+impl Scheduler for PrioGraphScheduler {
     /// Schedule transactions from the given `TransactionStateContainer` to be
     /// consumed by the worker threads. Returns summary of scheduling, or an
     /// error.
@@ -80,7 +84,7 @@ impl PrioGraphScheduler {
     /// This, combined with internal tracking of threads' in-flight transactions, allows
     /// for load-balancing while prioritizing scheduling transactions onto threads that will
     /// not cause conflicts in the near future.
-    pub(crate) fn schedule(
+    fn schedule(
         &mut self,
         container: &mut TransactionStateContainer,
         pre_graph_filter: impl Fn(&[&SanitizedTransaction], &mut [bool]),
@@ -300,7 +304,7 @@ impl PrioGraphScheduler {
 
     /// Receive completed batches of transactions without blocking.
     /// Returns (num_transactions, num_retryable_transactions) on success.
-    pub fn receive_completed(
+    fn receive_completed(
         &mut self,
         container: &mut TransactionStateContainer,
     ) -> Result<(usize, usize), SchedulerError> {
@@ -316,7 +320,9 @@ impl PrioGraphScheduler {
         }
         Ok((total_num_transactions, total_num_retryable))
     }
+}
 
+impl PrioGraphScheduler {
     /// Receive completed batches of transactions.
     /// Returns `Ok((num_transactions, num_retryable))` if a batch was received, `Ok((0, 0))` if no batch was received.
     fn try_receive_completed(
@@ -442,7 +448,7 @@ impl PrioGraphScheduler {
     ///
     /// Panics if the `thread_set` is empty. This should never happen, see comment
     /// on `ThreadAwareAccountLocks::try_lock_accounts`.
-    fn select_thread(
+    pub(crate) fn select_thread(
         thread_set: ThreadSet,
         batch_cus_per_thread: &[u64],
         in_flight_cus_per_thread: &[u64],
@@ -482,28 +488,15 @@ impl PrioGraphScheduler {
     }
 }
 
-/// Metrics from scheduling transactions.
-#[derive(Debug, PartialEq, Eq)]
-pub(crate) struct SchedulingSummary {
-    /// Number of transactions scheduled.
-    pub num_scheduled: usize,
-    /// Number of transactions that were not scheduled due to conflicts.
-    pub num_unschedulable: usize,
-    /// Number of transactions that were dropped due to filter.
-    pub num_filtered_out: usize,
-    /// Time spent filtering transactions
-    pub filter_time_us: u64,
-}
-
-struct Batches {
-    ids: Vec<Vec<TransactionId>>,
-    transactions: Vec<Vec<SanitizedTransaction>>,
-    max_ages: Vec<Vec<MaxAge>>,
-    total_cus: Vec<u64>,
+pub(crate) struct Batches {
+    pub ids: Vec<Vec<TransactionId>>,
+    pub transactions: Vec<Vec<SanitizedTransaction>>,
+    pub max_ages: Vec<Vec<MaxAge>>,
+    pub total_cus: Vec<u64>,
 }
 
 impl Batches {
-    fn new(num_threads: usize) -> Self {
+    pub(crate) fn new(num_threads: usize) -> Self {
         Self {
             ids: vec![Vec::with_capacity(TARGET_NUM_TRANSACTIONS_PER_BATCH); num_threads],
             transactions: vec![Vec::with_capacity(TARGET_NUM_TRANSACTIONS_PER_BATCH); num_threads],
@@ -512,7 +505,7 @@ impl Batches {
         }
     }
 
-    fn take_batch(
+    pub(crate) fn take_batch(
         &mut self,
         thread_id: ThreadId,
     ) -> (
@@ -540,15 +533,15 @@ impl Batches {
 }
 
 /// A transaction has been scheduled to a thread.
-struct TransactionSchedulingInfo {
-    thread_id: ThreadId,
-    transaction: SanitizedTransaction,
-    max_age: MaxAge,
-    cost: u64,
+pub(crate) struct TransactionSchedulingInfo {
+    pub thread_id: ThreadId,
+    pub transaction: SanitizedTransaction,
+    pub max_age: MaxAge,
+    pub cost: u64,
 }
 
 /// Error type for reasons a transaction could not be scheduled.
-enum TransactionSchedulingError {
+pub(crate) enum TransactionSchedulingError {
     /// Transaction was filtered out before locking.
     Filtered,
     /// Transaction cannot be scheduled due to conflicts, or