+++ core/src/banking_stage/transaction_scheduler/scheduler_controller.rs
@@ -3,7 +3,7 @@
 
 use {
     super::{
-        prio_graph_scheduler::PrioGraphScheduler,
+        scheduler::Scheduler,
         scheduler_error::SchedulerError,
         scheduler_metrics::{
             SchedulerCountMetrics, SchedulerLeaderDetectionMetrics, SchedulerTimingMetrics,
@@ -46,7 +46,11 @@ use {
 };
 
 /// Controls packet and transaction flow into scheduler, and scheduling execution.
-pub(crate) struct SchedulerController<T: LikeClusterInfo> {
+pub(crate) struct SchedulerController<C, S>
+where
+    C: LikeClusterInfo,
+    S: Scheduler,
+{
     /// Decision maker for determining what should be done with transactions.
     decision_maker: DecisionMaker,
     /// Packet/Transaction ingress.
@@ -58,7 +62,7 @@ pub(crate) struct SchedulerController<T: LikeClusterInfo> {
     /// Shared resource between `packet_receiver` and `scheduler`.
     container: TransactionStateContainer,
     /// State for scheduling and communicating with worker threads.
-    scheduler: PrioGraphScheduler,
+    scheduler: S,
     /// Metrics tracking time for leader bank detection.
     leader_detection_metrics: SchedulerLeaderDetectionMetrics,
     /// Metrics tracking counts on transactions in different states
@@ -70,17 +74,21 @@ pub(crate) struct SchedulerController<T: LikeClusterInfo> {
     /// Metric report handles for the worker threads.
     worker_metrics: Vec<Arc<ConsumeWorkerMetrics>>,
     /// State for forwarding packets to the leader, if enabled.
-    forwarder: Option<Forwarder<T>>,
+    forwarder: Option<Forwarder<C>>,
 }
 
-impl<T: LikeClusterInfo> SchedulerController<T> {
+impl<C, S> SchedulerController<C, S>
+where
+    C: LikeClusterInfo,
+    S: Scheduler,
+{
     pub fn new(
         decision_maker: DecisionMaker,
         packet_deserializer: PacketDeserializer,
         bank_forks: Arc<RwLock<BankForks>>,
-        scheduler: PrioGraphScheduler,
+        scheduler: S,
         worker_metrics: Vec<Arc<ConsumeWorkerMetrics>>,
-        forwarder: Option<Forwarder<T>>,
+        forwarder: Option<Forwarder<C>>,
     ) -> Self {
         Self {
             decision_maker,
@@ -714,8 +722,10 @@ mod tests {
         crate::{
             banking_stage::{
                 consumer::TARGET_NUM_TRANSACTIONS_PER_BATCH,
+                packet_deserializer::PacketDeserializer,
                 scheduler_messages::{ConsumeWork, FinishedConsumeWork, TransactionBatchId},
                 tests::create_slow_genesis_config,
+                transaction_scheduler::prio_graph_scheduler::PrioGraphScheduler,
             },
             banking_trace::BankingPacketBatch,
             sigverify::SigverifyTracerPacketStats,
@@ -758,7 +768,12 @@ mod tests {
         finished_consume_work_sender: Sender<FinishedConsumeWork>,
     }
 
-    fn create_test_frame(num_threads: usize) -> (TestFrame, SchedulerController<Arc<ClusterInfo>>) {
+    fn create_test_frame(
+        num_threads: usize,
+    ) -> (
+        TestFrame,
+        SchedulerController<Arc<ClusterInfo>, PrioGraphScheduler>,
+    ) {
         let GenesisConfigInfo {
             mut genesis_config,
             mint_keypair,
@@ -853,7 +868,7 @@ mod tests {
     // In the tests, the decision will not become stale, so it is more convenient
     // to receive first and then schedule.
     fn test_receive_then_schedule(
-        scheduler_controller: &mut SchedulerController<Arc<ClusterInfo>>,
+        scheduler_controller: &mut SchedulerController<Arc<ClusterInfo>, impl Scheduler>,
     ) {
         let decision = scheduler_controller
             .decision_maker